requirements.txt
langgraph
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
sqlite-utils
streamlit
python-dotenv
